islip: * compiler, interpreter, bytecode: added support for conditional (if-else) statements; elseif is on the way * compiler, parser: moved comment handling to parser * compiler: cleaned up the newline handling mess * makefile updates * compiler: further optimization development; need to make it context-sensitive so as not to fuck up good code
git-svn-id: https://subversion.assembla.com/svn/iqpk@14 ba71ee55-0f36-41dd-98d7-ab351bc57a59

+++ islip/bytecode.pas
@@ -150,7 +150,7 @@ type
     // single instruction with the argument
     islip_inst      = record
         inst        : byte;
-        arg         : size_t;
+        arg         : int;
     end;
     
     islip_bytecode  = array of islip_inst;
